refactor: add missing ng prefix to the angular module name
replace module name from 'ovh-angular-responsive-popover' to 'ngOvhResponsivePopover'

BREAKING CHANGE: module is now named as `ngOvhResponsivePopover`
