+++ package.json
@@ -1,22 +1,24 @@
 {
-  "name": "ovh-angular-responsive-popover",
+  "name": "@ovh-ux/ng-ovh-responsive-popover",
   "version": "5.0.0-alpha.2",
   "description": "Module is used to display a popover and adapt it to the screen dimension.",
   "keywords": [
     "angular"
   ],
-  "repository": "ovh-ux/ovh-angular-responsive-popover",
+  "repository": "ovh-ux/ng-ovh-responsive-popover",
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "files": [
     "dist"
   ],
-  "main": "./dist/cjs/ovh-angular-responsive-popover.js",
-  "browser": "./dist/umd/ovh-angular-responsive-popover.js",
+  "main": "./dist/cjs/index.js",
+  "browser": "./dist/umd/ng-ovh-responsive-popover.js",
   "scripts": {
     "build": "rollup -c",
-    "lint": "run-p lint:css lint:js lint:md",
+    "dev": "rollup -c --watch",
+    "lint": "run-p lint:html lint:css lint:js lint:md",
     "lint:css": "stylelint 'src/**/*.less' --fix",
+    "lint:html": "htmlhint 'src/**/*.html'",
     "lint:js": "eslint --quiet --fix --format=pretty src",
     "lint:md": "eslint --quiet --fix --format=pretty --ext .md .",
     "prepare": "rollup -c",
