+++ rollup.config.js
@@ -1,7 +1,7 @@
 import rollupConfig from '@ovh-ux/component-rollup-config';
 
 const config = rollupConfig({
-  input: './src/ovh-angular-responsive-popover.js',
+  input: 'src/index.js',
 });
 
 export default [
