+++ src/class/directive.js
