+++ src/class/ovh-angular-responsive-popover-class.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import directive from './directive';
+
+const moduleName = 'ngOvhResponsivePopoverClass';
+
+angular
+  .module(moduleName, [])
+  .directive('responsivePopoverClass', directive);
+
+export default moduleName;
