+++ src/popup/index.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import directive from './directive';
+
+const moduleName = 'ngOvhResponsivePopoverPopup';
+
+angular
+  .module(moduleName, [])
+  .directive('responsivePopoverPopup', directive);
+
+export default moduleName;
