+++ src/popup/template.html
